# openApi.yaml
openapi: 3.0.0
info:
  title: 'PasswordBroker API'
  description: ''
  version: v1
paths:
...
  /identity/api/users/search:
    get:
      tags:
        - Identity_UserController
      summary: 'Users list'
      operationId: ff3c8c656d0fbe284e83596b112d4361
      parameters:
        -
          name: q
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Identity_UsersSearchRequest_q'
        -
          name: perPage
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Identity_UsersSearchRequest_perPage'
        -
          name: page
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Identity_UsersSearchRequest_page'
        -
          name: entryGroupInclude
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Identity_UsersSearchRequest_entryGroupInclude'
        -
          name: entryGroupExclude
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Identity_UsersSearchRequest_entryGroupExclude'
      responses:
        '200':
          description: 'List of user with pagination'
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Common_Paginator' }
                  - { description: 'User set as data type', properties: { data: { type: array, items: { $ref: '#/components/schemas/Identity_User' } } }, type: object }
  /system/api/backups:
    get:
      tags:
        - System_BackupController
      summary: 'Backup entities list'
      operationId: 1cdd20acd2c39eb2bbc267e626cf450e
      parameters:
        -
          name: q
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/System_BackupSearchRequest_q'
        -
          name: perPage
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/System_BackupSearchRequest_perPage'
        -
          name: page
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/System_BackupSearchRequest_page'
      responses:
        '200':
          description: 'List of backups with pagination'
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Common_Paginator' }
                  - { description: 'Backup set as data type', properties: { data: { type: array, items: { $ref: '#/components/schemas/System_Backup' } } }, type: object }
    post:
      tags:
        - System_BackupController
      summary: 'Create new Backup'
      operationId: f6d7c03a1667c4ae9470922fdf58e1c0
      responses:
        '200':
          description: 'Backup entity was successfully created and put in the queue for executing'
  '/backups/{backup:backup_id}':
    get:
      tags:
        - System_BackupController
      summary: 'Show a Backup entity'
      operationId: 9e51625c7d80eac2fad5593b565bdfc9
      parameters:
        -
          name: 'backup:backup_id'
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/System_BackupId'
      responses:
        '200':
          $ref: '#/components/schemas/System_Backup'
components:
  schemas:
...
    Identity_UsersSearchRequest_q:
      type: string
      nullable: true
    Identity_UsersSearchRequest_entryGroupInclude:
      type: string
      format: uuid
      nullable: true
    Identity_UsersSearchRequest_entryGroupExclude:
      type: string
      format: uuid
      nullable: true
    Identity_UsersSearchRequest_perPage:
      type: integer
      minimum: 1
      example: 15
      nullable: true
    Identity_UsersSearchRequest_page:
      type: integer
      maximum: 100
      minimum: 1
      example: 1
      nullable: true
...
    Identity_Email:
      type: string
      format: email
    Identity_IsAdmin:
      type: boolean
    Identity_PublicKey:
      type: string
      format: binary
    Identity_UserId:
      type: string
      format: uuid
    Identity_UserName:
      type: string
    Identity_User:
      properties:
        user_id:
          $ref: '#/components/schemas/Identity_UserId'
        public_key:
          $ref: '#/components/schemas/Identity_PublicKey'
        is_admin:
          $ref: '#/components/schemas/Identity_IsAdmin'
        name:
          $ref: '#/components/schemas/Identity_UserName'
        email:
          $ref: '#/components/schemas/Identity_Email'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
      type: object
...
    System_BackupSearchRequest_q:
      type: string
      nullable: true
    System_BackupSearchRequest_perPage:
      type: integer
      minimum: 1
      example: 15
      nullable: true
    System_BackupSearchRequest_page:
      type: integer
      maximum: 100
      minimum: 1
      example: 1
      nullable: true
    System_BackupCreated:
      type: string
      format: date-time
    System_BackupDeleted:
      type: string
      format: date-time
      nullable: true
    System_BackupId:
      type: string
      format: uuid
    System_BackupPassword:
      type: string
      nullable: true
    System_BackupState:
      type: string
      enum:
        - await
        - creating
        - created
        - error
        - deleted
    System_ErrorMessage:
      type: string
      nullable: true
    System_FileName:
      type: string
    System_Size:
      type: integer
    System_Backup:
      properties:
        backup_id:
          $ref: '#/components/schemas/System_BackupId'
        file_name:
          $ref: '#/components/schemas/System_FileName'
        state:
          $ref: '#/components/schemas/System_BackupState'
        size:
          $ref: '#/components/schemas/System_Size'
        backup_created:
          $ref: '#/components/schemas/System_BackupCreated'
        backup_deleted:
          $ref: '#/components/schemas/System_BackupDeleted'
        error_message:
          $ref: '#/components/schemas/System_ErrorMessage'
        password:
          $ref: '#/components/schemas/System_BackupPassword'
      type: object
    Common_Paginator:
      properties:
        data: {  }
        current_page:
          type: integer
          minimum: 1
        last_page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
      type: object